Factor apart server into separate modules
Adds the following modules, factored apart from server:

- entry
- id
- path
