+++ src/main.rs
@@ -11,12 +11,15 @@ extern crate time;
 mod adapter;
 mod block;
 mod direntry;
+mod entry;
 mod error;
+mod id;
 mod lmdb_adapter;
 mod objectclass;
 mod objecthash;
 mod op;
 mod password;
+mod path;
 mod server;
 mod signature;
