+++ src/server.rs
@@ -1,15 +1,13 @@
-use std::{self, mem, str};
-use std::hash::{Hash, Hasher};
+use std::{self, str};
 use std::collections::HashMap;
 
 use ring::rand;
 
 use adapter::{Adapter, Transaction};
 use block::{Block, DigestAlgorithm};
-use direntry::DirEntry;
 use error::{Error, Result};
+use id::Id;
 use lmdb_adapter::LmdbAdapter;
-use objectclass::ObjectClass;
 use op::OpType;
 use password::{self, PasswordAlgorithm};
 use signature::{SignatureAlgorithm, KeyPair};
@@ -19,109 +17,8 @@ extern crate tempdir;
 
 const DEFAULT_GENESIS_MESSAGE: &'static str = "Initial block";
 
-#[derive(Debug, Eq, PartialEq, Copy, Clone)]
-pub struct Id(u64);
-
-#[derive(Debug, Eq, PartialEq)]
-pub struct Entry<'a> {
-    pub direntry: DirEntry<'a>,
-    pub objectclass: ObjectClass,
-}
-
-#[derive(Debug, Eq, PartialEq)]
-pub struct Path {
-    pub components: Vec<String>,
-}
-
 pub struct Server {
     adapter: LmdbAdapter,
-}
-
-// Ids are 64-bit integers in host-native byte order
-// LMDB has special optimizations for host-native integers as keys
-impl Id {
-    pub fn root() -> Id {
-        Id(0)
-    }
-
-    pub fn from_bytes(bytes: &[u8]) -> Result<Id> {
-        if bytes.len() != 8 {
-            return Err(Error::Parse);
-        }
-
-        let mut id = [0u8; 8];
-        id.copy_from_slice(&bytes[0..8]);
-
-        Ok(Id(unsafe { mem::transmute(id) }))
-    }
-
-    pub fn as_bytes(self) -> [u8; 8] {
-        let Id(id) = self;
-        unsafe { mem::transmute(id) }
-    }
-
-    pub fn next(self) -> Id {
-        let Id(id) = self;
-        Id(id + 1)
-    }
-}
-
-impl Path {
-    pub fn new(string: &str) -> Result<Path> {
-        let mut components: Vec<String> =
-            string.split("/").map(|component| String::from(component)).collect();
-
-        if components.is_empty() {
-            return Err(Error::PathInvalid);
-        }
-
-        let prefix = components.remove(0);
-
-        // Does the path start with something other than "/"?
-        if !prefix.is_empty() {
-            return Err(Error::PathInvalid);
-        }
-
-        Ok(Path { components: components })
-    }
-
-    pub fn to_string(&self) -> String {
-        let mut result = String::new();
-
-        for component in self.components.clone() {
-            result.push_str("/");
-            result.push_str(&component);
-        }
-
-        result
-    }
-
-    pub fn parent(&self) -> Path {
-        if self.is_root() {
-            return Path { components: vec![String::from("")] };
-        }
-
-        let mut parent_components = self.components.clone();
-        parent_components.pop();
-
-        Path { components: parent_components }
-    }
-
-    pub fn name(&self) -> String {
-        self.components.last().unwrap().clone()
-    }
-
-    pub fn is_root(&self) -> bool {
-        self.components.len() == 1 && self.components[0] == ""
-    }
-}
-
-impl Hash for Path {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        for component in &self.components {
-            component.hash(state);
-        }
-    }
 }
 
 impl Server {
